@@ -16,21 +16,22 @@
 
 package de.schosinsky.sample;
 
-import com.blazebit.persistence.Criteria;
-import com.blazebit.persistence.CriteriaBuilderFactory;
-import com.blazebit.persistence.spi.CriteriaBuilderConfiguration;
-import java.util.function.Consumer;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityTransaction;
-import javax.persistence.Persistence;
-import de.schosinsky.model.Cat;
-import de.schosinsky.model.Person;
+import java.util.function.BiConsumer;
+
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityTransaction;
+import jakarta.persistence.Persistence;
+
 import org.junit.After;
 import org.junit.Before;
 
+import com.blazebit.persistence.Criteria;
+import com.blazebit.persistence.CriteriaBuilderFactory;
+import com.blazebit.persistence.spi.CriteriaBuilderConfiguration;
+
 public abstract class AbstractSampleTest {
-    
+
     protected EntityManagerFactory emf;
     protected CriteriaBuilderFactory cbf;
 
@@ -39,56 +40,19 @@ public abstract class AbstractSampleTest {
         emf = Persistence.createEntityManagerFactory("default");
         CriteriaBuilderConfiguration config = Criteria.getDefault();
         cbf = config.createCriteriaBuilderFactory(emf);
-        
-        transactional(em -> {
-            Person p1 = new Person("P1");
-            Person p2 = new Person("P2");
-            Person p3 = new Person("P3");
-            em.persist(p1);
-            em.persist(p2);
-            em.persist(p3);
-            
-            Cat c1 = new Cat("C1", 1, p2);
-            Cat c2 = new Cat("C2", 2, p2);
-            Cat c3 = new Cat("C3", 4, p2);
-            
-            Cat c4 = new Cat("C4", 6, p3);
-            
-            Cat c5 = new Cat("C5", 8, null);
-            Cat c6 = new Cat("C6", 7, null);
-            
-            em.persist(c1);
-            em.persist(c2);
-            em.persist(c3);
-            em.persist(c4);
-            em.persist(c5);
-            em.persist(c6);
-            
-            c1.setMother(c3);
-            c3.getKittens().add(c1);
-            
-            c1.setFather(c5);
-            c5.getKittens().add(c1);
-            
-            c2.setMother(c3);
-            c3.getKittens().add(c2);
-            
-            c2.setFather(c6);
-            c6.getKittens().add(c2);
-            
-            c4.setFather(c6);
-            c6.getKittens().add(c4);
-        });
     }
-    
-    protected void transactional(Consumer<EntityManager> consumer) {
+
+    protected void transactional(BiConsumer<EntityManager, CriteriaBuilderFactory> consumer) {
         EntityManager em = emf.createEntityManager();
         EntityTransaction tx = em.getTransaction();
         boolean success = false;
-        
+
+        var criteria = Criteria.getDefault();
+        var criteriaBuilderFactory = criteria.createCriteriaBuilderFactory(emf);
+
         try {
             tx.begin();
-            consumer.accept(em);
+            consumer.accept(em, criteriaBuilderFactory);
             success = true;
         } finally {
             try {
